Renames header
